requirements.txt
PyQt6
mss
pynput
psutil
Pillow
pygetwindow

pywin32
